Added the classes CCitiesTable, CCitiesAccessor and the structure City which manage a connection with a table in a database. Added the CLog class which stores information about application errors in file named log.txt and creates log directory named log.
CCitiesTable open a connection with MSSQL server, opens a connection and run a query that load all records from the table Cities. The class uses the OLE DB driver. Eventual errors are handled by displaying Messages in the GUI, using the CMessage class and logging error information in ProjectRootFoler/log/log.txt using the CLog class. Closing the rowset, session and the connection with the database is done within the class.

The structure City mimics the structure of a record from the database table. The data in the rowset in the CCitiesTable class consists of a collection of instances of this structure.

The class CCitiesAccessor maps the data from the record of the database table to the different fields of the struct City. It consists of 2 accessors. The first one holds only the id, the second one holds everything else. When reading from the database, both accessors are used, when writing to the databse, only the second accsessor is used.

The class CMessage displays dialog based message that informs the user that some event occured, mainly errors.

The class CLog is used by CMessage to store more detailed error information in a text file in ProjectRootFolder/lo/log.txt which is for the developers and not for the users. Every time a new instance of the application is started and the instance have called the logger, the logfile is truncated, which means it is cleared and only the latest errors are shown.