CCitiesTable's public methods are changed. CDBConnection is changed. The database connection is in singleton, it opens the first time a reference to the connection is created and is closed when the application terminates. CLog is changed.
The CCitiesTable's loadCities is renamed to SelectAll. It selects all records from the database table Cities. The method SelectWhereId is added to CCitiesTable. It selects just one record with the specified id. If no records or more than one is found is considered an error.

The CDBConnection is in a singleton which opens and closes the database connection only once. The lifetime of the connection is from the first time it is needed to the end of the application.

The CLog's overloaded methods LogMessage are now static and call the singleton's instance within.

The method getFileNameFromPath is moved from CMessage to CLog.