Added CCitiesData. Added ordered_map to the CCitiesDoc. The session is no longer responsability of the CCitiesTable,  its responsability to the CCitiesData. CCitiesView::OnInitialUpdate is changed.
The CitiesData is a layer that takes care of the business logic and stays between the document and the table layer. The method SelectAll() is added. It calls the CCitiesTable::SelectAll() and gives it a reference of a map from the CCitiesDoc.

The map in the CCitiesDoc will be the main container for the data from the database table Cities. It will sort the records by the alphabetical order of the combination of the city name and area.

The session is created in the CCitiesData and is passed to CCitiesTable methods as an argument. The reason is that if the methods of the data layer requires data from more than one table at the same time, one session with one transaction should be used for all required tables. The session must be closed in the data layer.

The CCitiesView now visualizes the data from the CCitiesDoc. The method OnInitialUpdate() is changed. Functionalities like configurationg the look of the list control of the view as well as populating the list control are added in the OnInitialUpdate();