Added CCitiesView. CCitiesDoc is updated. CMessage and CLog are updated. Dialogs for the data of a selected record and for message to the user are added.
Added CitiesDoc methods for selecting, updating, inserting and deleting records from the database and the document's container which is std::ordered_map.

Added CitiesView methods for initializing the view, handling view, add, change and remove record operations, opening context menu with those options. Those methods handle the refreshing the outdated records with which the user interacts.

Added dialog for the displaying the data of a selected city. The dialog have 4 different modes, depending of the operation that the user want to execute.

Added dialog for messages that notifies the user for errors, warnings, notifications or if they need to choose action. The dialog have 4 modes depending of the situation.

The CLog is changed. Now it shows all data in visually cleaner and ordered manner. The name of the metod is added in the log.